Add integration tests
I wrote a few integration test.
And also an task in the Rakefile to test this. It took me hours to find
out how I can start and end the webserver, while running the tests in
the middle. So please forgive me, if it's not the best solution, I'm
glad it is running.