Add handler for manage different accempt types
The default is html/text.
Json and xml are supported too. Maybe I have to reimplement the to_json and the
to_xml functionality because the style of the result is not exaclty what I
wanted.
The slim template works :)

Next step: use the result handler and hide the function somewhere else for
better overview.